timetable screen rework process

+++ app/src/main/java/com/example/deathnote/presentation/ui/screen/main_screen/components/statistics_screen_ui/Stats1_M.kt
@@ -33,7 +33,7 @@ fun Stats1_M(
     subject: Subject,
     respectfulAbsences: Int = 0,
     absences: Int = 0,
-    absencePercent: Int = 0
+    absencePercent: Float = 0f
 ) {
 
     Row(
